+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Added
 * Added GenericMixerManager
+* Added unit testing for Floater
 
 Changed
 * AsyncCollection, optimize catch exception